chore: restore coverage data
